fix: Item dto 수정 #51

+++ monew/src/main/java/com/codeit/team2/monew/module/domain/useractivity/dto/CommentLikeItem.java
@@ -12,7 +12,7 @@ public record CommentLikeItem(
     UUID commentUserId,
     String commentUserNickname,
     String commentContent,
-    Integer commentLikeCount,
+    Long commentLikeCount,
     Instant commentCreatedAt
 ) {
